pdfpres: Encode range of preview windows in instance name.
Each showpdf window will have an instance name like "a_b_c" where "a" is
the minimal offset, "b" the maximum offset and "c" the offset for this
window.

Say you have 3 preview windows, one for the previous slide, one for the
next slide and one for the current slide. Then your showpdf windows will
have the following instance names:

	-1_1_-1    (previous slide)
	-1_1_0     (current slide)
	-1_1_1     (next slide)

This will make it easier for the window manager to deal with different
configurations.